Move workout item into separate component and more clean up
